Changed most floats to double for the functions for camera lib,direction lib and event detect. Also created a new folder/Project with 200ksps sample rate pr ADC.

+++ main_arduino/lib/direction/direction.h
@@ -1,6 +1,6 @@
 #include <math.h>
 #include <Camera.h>
 
-float yy1(float x, float da, float l);
+double yy1(double x, double da, double l);
 
-angleSet direction(float ta, float tb, float tc); 
+angleSet direction(double ta, double tb, double tc);
